Updated Signup.vue & create GoogleLogin component
