+++ src/components/Login.vue
@@ -28,15 +28,7 @@
         <span id="textDivider">or sign in with</span>
       </p>
 
-      <button @click="socialLogin" class="googlebtn">
-        <img
-          alt="Google Logo"
-          src="../assets/google-logo.png"
-          width="24px"
-          height="24px"
-        />
-        <span class="googelbtntxt">Google</span>
-      </button>
+      <google-login></google-login>
     </div>
 
     <p>
@@ -47,8 +39,8 @@
 </template>
 
 <script>
-import firebase from "firebase";
-import database from "../firebase.js";
+import GoogleLogin from "./GoogleLogin";
+
 export default {
   name: "login",
   data() {
@@ -57,91 +49,10 @@ export default {
       password: "",
     };
   },
+  components: {
+    GoogleLogin,
+  },
   methods: {
-    socialLogin: function () {
-      const provider = new firebase.auth.GoogleAuthProvider();
-      firebase
-        .auth()
-        .signInWithPopup(provider)
-        .then((result) => {
-          console.log(result.user);
-
-          database
-            .firestore()
-            .collection("users")
-            .doc(result.user.uid)
-            .get()
-            .then(async (doc) => {
-              console.log("Exists: " + doc.exists);
-              if (doc.exists) {
-                console.log("user exists");
-                this.$router.push({ name: "Home" });
-              } else {
-                console.log("user do not exist");
-                this.createAccount(
-                  result.user.displayName,
-                  result.user.email,
-                  result.user.photoURL,
-                  result.user.uid
-                );
-                this.$router.push({ name: "SetPassword" });
-              }
-            })
-            .catch((error) => {
-              console.log("Error getting document:", error);
-            });
-        })
-        .catch((err) => {
-          alert("error " + err.message);
-        });
-    },
-    createAccount: function (name, email, photoUrl, uid) {
-      console.log("Enter create ACCOUNT ");
-      console.log("name: " + name);
-      console.log("email: " + email);
-      console.log("photoUrl: " + photoUrl);
-      console.log("uid: " + uid);
-      database
-        .firestore()
-        .collection("users")
-        .doc(uid)
-        .set({
-          defaulyPaymentMethod: "",
-          email: email,
-          name: name,
-          phone: "",
-          pools: [],
-          profilePhoto: photoUrl,
-          pw: "",
-        })
-        .then(() => {
-          location.reload();
-        })
-        .catch((error) => {
-          console.error("Error writing document: ", error);
-        });
-    },
-    async checkUser(uid) {
-      var exist;
-      await database
-        .firestore()
-        .collection("users")
-        .doc(uid)
-        .get()
-        .then(async (doc) => {
-          console.log("Exists?" + doc.exists);
-          if (doc.exists) {
-            exist = true;
-          } else {
-            exist = false;
-          }
-        })
-        .catch((error) => {
-          console.log("Error getting document:", error);
-        });
-      console.log("asdasdasdasd" + exist);
-      return exist;
-    },
     remember_me: function () {
       console.log("Remember me");
     },
